requirements.txt
cryptography
discord.py
requests
mysql-connector-python
psutil
pymongo
dnspython